requirements.txt
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
Django==4.0.6
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.0
filelock==3.4.2
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
linecache2==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
traceback2==1.4.0
tzdata==2022.1
unittest2==1.1.0
uritemplate==4.1.1
urllib3==1.26.10
uuid==1.30
virtualenv==20.13.0
